requirements.txt
streamlit>=1.28.0
pydantic>=2.5.2
litellm>=1.9.6
python-dotenv>=1.0.0
watchdog>=3.0.0
openai>=1.3.5
numpy>=1.24.4
pandas>=2.0.3
pillow>=10.1.0
crawl4ai>=0.2.8
streamlit-extras>=0.3.5
playwright>=1.40.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
nest_asyncio>=1.5.6
streamlit-tags>=1.2.8
openpyxl


